pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "acadlabs-cli"
version = "0.1.0"
description = "AI-powered coding assistant CLI"
authors = [{name = "Acadlabs Team"}]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
acadlabs = "acadlabs_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/acadlabs_cli"]